refactor(polynomials): partition core multivariate contracts
