+++ src/jacobian/math/polynomials/multivariate/_models.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass
 from fractions import Fraction
 from functools import reduce
 from math import comb, factorial, gcd, lcm
-from typing import Annotated, Any, Literal, Self
+from typing import Any, Literal, Self
 
 from pydantic import Field, StrictInt, model_validator
 from pydantic_core import PydanticCustomError
@@ -28,10 +28,6 @@ _MAX_MULTIVARIATE_TERMS = 512
 _MAX_MULTIVARIATE_EXPONENT = 64
 _MAX_MULTIVARIATE_COEFFICIENT_DIGITS = 256
 _MAX_ELIMINATION_DEGREE_SUM = 64
-# The result converter rejects sparse outputs above this size.  The request
-# validator uses the same bound to reject large possible supports before
-# SymPy expands the Sylvester determinant.
-_MAX_RESULTANT_TERMS = 1_024
 # Public output-term budget for one converted irreducible factor.  The
 # operation converter uses this same bound; keeping it here lets the result
 # validator reproduce the kernel's exact exceedance decision.
@@ -51,9 +47,6 @@ _SUBRESULTANT_BACKEND_PASS_COUNT = 5
 _MAX_SUBRESULTANT_MEMBER_COUNT = _MAX_ELIMINATION_DEGREE_SUM // 2 + 2
 _SUBRESULTANT_RESULT_TERM_MULTIPLIER = 4
 
-MonomialOrder = Literal["lex", "grlex", "grevlex"]
-"""Declared monomial order for multivariate polynomial division."""
-
 _MULTIVARIATE_MIN_VARIABLES = 2
 
 
@@ -69,50 +62,6 @@ def _validate_multivariate_pair(
         )
     if left.variables != right.variables:
         raise _validation_error("both polynomials must use the same ordered variables")
-
-
-def _resultant_support_bound(
-    left: RationalPolynomial,
-    right: RationalPolynomial,
-    elimination_index: int,
-) -> int:
-    """Bound the number of possible monomials in a multivariate resultant.
-
-    If ``f`` and ``g`` have elimination degrees ``m`` and ``n``, and total
-    remaining-variable degrees ``d_f`` and ``d_g``, every resultant monomial
-    has total degree at most ``n*d_f + m*d_g``.  The returned binomial is the
-    number of monomials up to that degree in the remaining variables.
-    """
-
-    remaining_variable_count = len(left.variables) - 1
-    if remaining_variable_count == 0:
-        return 1
-
-    def degree(polynomial: RationalPolynomial, *, in_remaining: bool) -> int:
-        return max(
-            (
-                sum(
-                    exponent
-                    for index, exponent in enumerate(term.exponents)
-                    if (index != elimination_index) == in_remaining
-                )
-                for term in polynomial.polynomial.terms
-            ),
-            default=0,
-        )
-
-    left_elimination_degree = degree(left, in_remaining=False)
-    right_elimination_degree = degree(right, in_remaining=False)
-    left_remaining_degree = degree(left, in_remaining=True)
-    right_remaining_degree = degree(right, in_remaining=True)
-    resultant_degree_bound = (
-        right_elimination_degree * left_remaining_degree
-        + left_elimination_degree * right_remaining_degree
-    )
-    return comb(
-        resultant_degree_bound + remaining_variable_count,
-        remaining_variable_count,
-    )
 
 
 def _degree_in_variable(polynomial: RationalPolynomial, variable_index: int) -> int:
@@ -340,160 +289,6 @@ def _subresultant_envelope(
             _SUBRESULTANT_RESULT_TERM_MULTIPLIER * aggregate_terms * coefficient_bits
         ),
     )
-
-
-class MultivariateGcdRequest(StrictModel):
-    """Two multivariate polynomials in ``QQ[x_1, ..., x_n]``."""
-
-    left: RationalPolynomial
-    right: RationalPolynomial
-
-    @model_validator(mode="after")
-    def require_multivariate_ring(self) -> Self:
-        _validate_multivariate_pair(self.left, self.right)
-        for polynomial in (self.left, self.right):
-            require_polynomial_budget(
-                polynomial,
-                maximum_terms=_MAX_MULTIVARIATE_TERMS,
-                maximum_exponent=_MAX_MULTIVARIATE_EXPONENT,
-                maximum_coefficient_digits=_MAX_MULTIVARIATE_COEFFICIENT_DIGITS,
-            )
-        return self
-
-
-class MultivariateGcdResult(StrictModel):
-    gcd: RationalPolynomial
-    convention: Literal["MONIC_ASSOCIATE"] = "MONIC_ASSOCIATE"
-
-
-class MultivariateDivisionRequest(StrictModel):
-    """Divide one multivariate polynomial by another under a declared monomial order."""
-
-    left: RationalPolynomial
-    right: RationalPolynomial
-    monomial_order: MonomialOrder = "lex"
-
-    @model_validator(mode="after")
-    def require_multivariate_ring(self) -> Self:
-        _validate_multivariate_pair(self.left, self.right)
-        if not self.right.polynomial.terms:
-            raise _validation_error("divisor polynomial must be nonzero")
-        for polynomial in (self.left, self.right):
-            require_polynomial_budget(
-                polynomial,
-                maximum_terms=_MAX_MULTIVARIATE_TERMS,
-                maximum_exponent=_MAX_MULTIVARIATE_EXPONENT,
-                maximum_coefficient_digits=_MAX_MULTIVARIATE_COEFFICIENT_DIGITS,
-            )
-        return self
-
-
-class MultivariateDivisionResult(StrictModel):
-    quotient: RationalPolynomial
-    remainder: RationalPolynomial
-    monomial_order: MonomialOrder
-    convention: Literal["EXACT_DIVISION_REMAINDER"] = "EXACT_DIVISION_REMAINDER"
-
-
-class MultivariateResultantRequest(StrictModel):
-    """Compute a bounded resultant with respect to one variable.
-
-    The eliminated-variable domain follows the Sylvester determinant exactly,
-    including its degenerate rows: an input that is a nonzero constant in the
-    elimination variable contributes the standard power rule
-    ``Res_x(f, c) = c ^ deg_x(f)`` (and symmetrically ``Res_x(c, g) =
-    c ^ deg_x(g)``), two inputs both constant in the eliminated variable give
-    the empty-determinant value 1, and a zero input gives 0.  The request
-    rejects inputs whose degree envelope can produce more terms than the
-    exact sparse result contract can represent.
-    """
-
-    left: RationalPolynomial
-    right: RationalPolynomial
-    elimination_variable: PolynomialVariable = Field(
-        description="Variable eliminated by the Sylvester resultant.",
-    )
-
-    @model_validator(mode="after")
-    def require_multivariate_ring(self) -> Self:
-        _validate_multivariate_pair(self.left, self.right)
-        if self.elimination_variable not in self.left.variables:
-            raise _validation_error(
-                "elimination variable must belong to the declared ring"
-            )
-        for polynomial in (self.left, self.right):
-            require_polynomial_budget(
-                polynomial,
-                maximum_terms=_MAX_MULTIVARIATE_TERMS,
-                maximum_exponent=_MAX_MULTIVARIATE_EXPONENT,
-                maximum_coefficient_digits=_MAX_MULTIVARIATE_COEFFICIENT_DIGITS,
-            )
-        variable_index = self.left.variables.index(self.elimination_variable)
-        degree_sum = max(
-            (term.exponents[variable_index] for term in self.left.polynomial.terms),
-            default=0,
-        ) + max(
-            (term.exponents[variable_index] for term in self.right.polynomial.terms),
-            default=0,
-        )
-        if degree_sum > _MAX_ELIMINATION_DEGREE_SUM:
-            raise _validation_error("Sylvester degree exceeds the resultant budget")
-        if (
-            _resultant_support_bound(self.left, self.right, variable_index)
-            > _MAX_RESULTANT_TERMS
-        ):
-            raise _validation_error("resultant output exceeds the term budget")
-        return self
-
-
-class MultivariateScalarValue(StrictModel):
-    kind: Literal["SCALAR"] = "SCALAR"
-    value: CanonicalRational
-
-
-class MultivariatePolynomialValue(StrictModel):
-    kind: Literal["POLYNOMIAL"] = "POLYNOMIAL"
-    value: RationalPolynomial
-
-
-MultivariateInvariantValue = Annotated[
-    MultivariateScalarValue | MultivariatePolynomialValue,
-    Field(discriminator="kind"),
-]
-
-
-class MultivariateResultantResult(StrictModel):
-    """The exact resultant bound to its source pair.
-
-    Retains both source polynomials and the eliminated variable so
-    validation replays the exact Sylvester determinant instead of trusting
-    an independently authored value.  The replay runs inside the same
-    admitted degree envelope as the producer.
-    """
-
-    left: RationalPolynomial
-    right: RationalPolynomial
-    elimination_variable: PolynomialVariable
-    resultant: MultivariateInvariantValue
-    convention: Literal["SYLVESTER_DETERMINANT"] = "SYLVESTER_DETERMINANT"
-
-    @model_validator(mode="after")
-    def require_source_bound(self) -> Self:
-        from jacobian.math.polynomials.multivariate._operations import (
-            _sylvester_resultant_value,
-        )
-
-        request = MultivariateResultantRequest(
-            left=self.left,
-            right=self.right,
-            elimination_variable=self.elimination_variable,
-        )
-        if self.resultant != _sylvester_resultant_value(request):
-            raise _validation_error(
-                "resultant must equal the Sylvester determinant of the "
-                "retained source polynomials"
-            )
-        return self
 
 
 class MultivariateSubresultantSequenceRequest(StrictModel):
@@ -1230,20 +1025,10 @@ def _verify_exact_reconstruction(
 
 
 __all__ = [
-    "MonomialOrder",
-    "MultivariateDivisionRequest",
-    "MultivariateDivisionResult",
     "MultivariateFactorRequest",
     "MultivariateFactorResult",
-    "MultivariateGcdRequest",
-    "MultivariateGcdResult",
-    "MultivariateInvariantValue",
     "MultivariateIrreducibleFactor",
-    "MultivariatePolynomialValue",
     "MultivariatePrincipalSubresultantCoefficient",
-    "MultivariateResultantRequest",
-    "MultivariateResultantResult",
-    "MultivariateScalarValue",
     "MultivariateSubresultantMember",
     "MultivariateSubresultantSequenceRequest",
     "MultivariateSubresultantSequenceResult",
